requirements.txt
dj-database-url==0.5.0
Django==2.1
django-heroku==0.3.1
pytz==2018.5
whitenoise==4.0
django-crispy-forms==1.7.2
django-mathfilters==0.4.0
django-filter==2.0.0
